sepearte a component of general Entry
